requirements.txt
pytest==8.0.2
selene==2.0.0rc9
allure-pytest==2.13.4
Appium-Python-Client==4.0.0
